@@ -3,31 +3,31 @@
   "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "auth0-js": "^9.8.2",
-    "auth0-lock": "^11.12.1",
-    "axios": "^0.18.0",
-    "bootstrap": "^4.2.1",
-    "jquery": "^3.3.1",
-    "moment": "^2.23.0",
+    "auth0-js": "^9.11.3",
+    "auth0-lock": "^11.18.1",
+    "axios": "^0.18.1",
+    "bootstrap": "^4.4.1",
+    "jquery": "^3.4.1",
+    "moment": "^2.24.0",
     "pdfobject": "^2.1.1",
-    "react": "^16.7.0",
-    "react-csv": "^1.0.19",
-    "react-dom": "^16.7.0",
-    "react-dropzone": "^8.0.3",
-    "react-ga": "^2.5.6",
-    "react-modal-video": "^1.2.1",
-    "react-redux": "^6.0.0",
-    "react-responsive": "^6.0.1",
-    "react-router-dom": "^4.3.1",
-    "react-scripts": "^2.1.3",
-    "react-scroll": "^1.7.10",
-    "react-spinners": "^0.5.1",
+    "react": "^16.12.0",
+    "react-csv": "^1.1.2",
+    "react-dom": "^16.12.0",
+    "react-dropzone": "^8.2.0",
+    "react-ga": "^2.7.0",
+    "react-modal-video": "^1.2.3",
+    "react-redux": "^7.1.3",
+    "react-responsive": "^6.1.2",
+    "react-router-dom": "^5.1.2",
+    "react-scripts": "^2.1.8",
+    "react-scroll": "^1.7.14",
+    "react-spinners": "^0.5.12",
     "react-stripe-checkout": "^2.6.3",
-    "reactstrap": "^7.0.2",
-    "redux": "^4.0.1",
+     "reactstrap": "^7.0.2",
+    "redux": "^4.0.4",
     "redux-thunk": "^2.3.0",
     "request": "^2.88.0",
-    "styled-components": "^4.1.3"
+    "styled-components": "^4.4.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -36,8 +36,8 @@
     "eject": "react-scripts eject"
   },
   "resolutions": {
-    "react": "16.7.0",
-    "react-dom": "16.7.0"
+    "react": "16.12.0",
+    "react-dom": "16.12.0"
   },
   "eslintConfig": {
     "extends": "react-app"
