@@ -1,4 +1,7 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
+import { useHistory } from 'react-router-dom';
+import { useDispatch } from 'react-redux';
+import { getProof, updateLoading } from '../../actions/envelopes';
 import {
   DocumentContainer,
   DocumentSubject,
@@ -16,71 +19,65 @@ import PDFModal from './PDFModal';
 import VerifyModal from './VerifyModal';
 import moment from 'moment';
 
-class Document extends React.Component {
-  constructor(props) {
-    super(props);
+const Document = props => {
+  const [document, setDocument] = useState(null);
 
-    this.state = {
-      modalLink: false,
-      modalVerify: false,
-      modalPdf: false,
-      document: null,
-    };
-  }
+  const [modalLink, setModalLink] = useState(false);
+  const [modalPdf, setModalPdf] = useState(false);
+  const [modalVerify, setModalVerify] = useState(false);
 
-  componentDidMount() {
-    if (this.props.doc.loading) {
-      this.checkLoading();
+  const dispatch = useDispatch();
+  const history = useHistory();
+
+  useEffect(() => {
+    let intervalId;
+
+    if (props.doc.loading) {
+      intervalId = checkLoading();
     }
-  }
 
-  componentWillUnmount() {
-    clearInterval(this.interval);
-  }
+    if (intervalId) return () => clearInterval(intervalId);
+  }, []);
 
-  getProof = () => {
-    this.props.getProof(this.props.doc.id);
-    this.toggleVerifyPay();
+  useEffect(() => {
+    if (document) {
+      togglePdfModal();
+    }
+  }, [document]);
+
+  const getProofA = () => {
+    dispatch(getProof(props.doc.id));
+    toggleVerifyPay();
   };
 
-  getDocuments = () => {
-    const envelope_id = this.props.doc.envelope_id;
-    if (process.env.REACT_APP_DOCUMENTS) {
-      axios
-        .get(`${process.env.REACT_APP_DOCUMENTS}/${envelope_id}`)
-        .then(res => {
-          const data = Buffer.from(res.data.document.data).toString();
-          this.setState({ document: data }, () => {
-            this.togglePdfModal();
-          });
-        })
-        .catch(err => console.log(err));
+  const getDocument = () => {
+    if (document) {
+      togglePdfModal();
     } else {
-      axios
-        .get(`http://localhost:9000/documents/${envelope_id}`)
-        .then(res => {
-          this.setState({ document: res.data.document }, () => {
-            this.togglePdfModal();
-          });
-        })
-        .catch(err => console.log(err));
+      const envelope_id = props.doc.envelope_id;
+      if (process.env.REACT_APP_DOCUMENTS) {
+        axios
+          .get(`${process.env.REACT_APP_DOCUMENTS}/${envelope_id}`)
+          .then(res => {
+            const data = Buffer.from(res.data.document.data).toString();
+            setDocument(data);
+          })
+          .catch(err => console.log(err));
+      } else {
+        axios
+          .get(`http://localhost:9000/documents/${envelope_id}`)
+          .then(res => setDocument(res.data.document))
+          .catch(err => console.log(err));
+      }
     }
   };
 
-  toggleLinkModal = () => {
-    this.setState({ modalLink: !this.state.modalLink });
-  };
+  const toggleLinkModal = () => setModalLink(!modalLink);
+  const togglePdfModal = () => setModalPdf(!modalPdf);
+  const toggleVerifyPay = () => setModalVerify(!modalVerify);
 
-  toggleVerifyPay = () => {
-    this.setState({ modalVerify: !this.state.modalVerify });
-  };
-
-  togglePdfModal = () => {
-    this.setState({ modalPdf: !this.state.modalPdf });
-  };
-
-  checkLoading = () => {
-    this.interval = setInterval(() => {
+  const checkLoading = () => {
+    const interval = setInterval(() => {
       let host;
       if (process.env.REACT_APP_CHAINPOINT) {
         host = process.env.REACT_APP_CHAINPOINT;
@@ -88,104 +85,98 @@ class Document extends React.Component {
         host = 'http://localhost:9000/chainpoint';
       }
 
-      const promise = axios.get(`${host}/${this.props.doc.id}/loading`);
+      const promise = axios.get(`${host}/${props.doc.id}/loading`);
 
       promise
         .then(res => {
           if (!res.data.loading) {
-            this.props.updateLoading(this.props.doc.id, res.data);
-            this.clearInterval(this.interval);
+            dispatch(updateLoading(props.doc.id, res.data));
+            clearInterval(interval);
           }
         })
-        .catch(() => clearInterval(this.interval));
+        .catch(() => clearInterval(interval));
     }, 5000);
+
+    return interval;
   };
 
-  render() {
-    // Removes "Please DocuSign: " from file title
-    const docuSignSubject = this.props.doc.subject;
-    const cutDocSignTitle = docuSignSubject.replace(/Please DocuSign: /, '');
+  // Removes "Please DocuSign: " from file title
+  const docuSignSubject = props.doc.subject;
+  const cutDocSignTitle = docuSignSubject.replace(/Please DocuSign: /, '');
 
-    // Timestamp of BTC proof
-    let timestamp;
+  // Timestamp of BTC proof
+  let timestamp;
 
-    if (this.props.doc.verified) {
-      const verified_proof = JSON.parse(this.props.doc.verified_proof);
-      timestamp = verified_proof.verifiedAt;
-    }
-
-    return (
-      <DocumentContainer>
-        <LinkModal
-          doc={this.props.doc}
-          toggle={this.toggleLinkModal}
-          isOpen={this.state.modalLink}
-        />
-
-        <VerifyModal
-          toggle={this.toggleVerifyPay}
-          isOpen={this.state.modalVerify}
-          getProof={this.getProof}
-        />
-
-        <PDFModal
-          doc={this.props.doc}
-          toggle={this.togglePdfModal}
-          isOpen={this.state.modalPdf}
-          document={this.state.document}
-        />
-
-        {this.props.doc.status === 'completed' &&
-        !this.props.doc.verified &&
-        !this.props.doc.waiting ? (
-          <DocumentProof
-            onClick={
-              this.props.user.credits <= 0
-                ? () => this.props.history.push('/buy')
-                : this.toggleVerifyPay
-            }
-          >
-            {this.props.doc.loading && !this.props.doc.error ? (
-              <LoadingContainer>
-                <BeatLoader color={'black'} sizeUnit={'px'} size={10} />
-              </LoadingContainer>
-            ) : this.props.doc.error ? (
-              'Error'
-            ) : (
-              'Click to Proof'
-            )}
-          </DocumentProof>
-        ) : (
-          <DocumentProof
-            onClick={this.props.doc.verified ? this.toggleLinkModal : null}
-          >
-            {this.props.doc.verified
-              ? 'See Proof'
-              : this.props.doc.waiting
-              ? 'Waiting...'
-              : 'Not signed'}
-          </DocumentProof>
-        )}
-        <ProofDocTextContainer>
-          <DocumentSubject onClick={this.getDocuments}>
-            {cutDocSignTitle} <PreviewIcon className="fas fa-search" />
-          </DocumentSubject>
-          {/* Returns proofed timestamp if exists */}
-          {timestamp !== undefined ? (
-            <TimestampContainer>
-              Proofed
-              <Timestamp>
-                {moment
-                  .utc(timestamp)
-                  .local()
-                  .format('D MMM YYYY hh:mma')}
-              </Timestamp>
-            </TimestampContainer>
-          ) : null}
-        </ProofDocTextContainer>
-      </DocumentContainer>
-    );
+  if (props.doc.verified) {
+    const verified_proof = JSON.parse(props.doc.verified_proof);
+    timestamp = verified_proof.verifiedAt;
   }
-}
+
+  return (
+    <DocumentContainer>
+      <LinkModal doc={props.doc} toggle={toggleLinkModal} isOpen={modalLink} />
+
+      <VerifyModal
+        toggle={toggleVerifyPay}
+        isOpen={modalVerify}
+        getProof={getProofA}
+      />
+
+      <PDFModal
+        doc={props.doc}
+        toggle={togglePdfModal}
+        isOpen={modalPdf}
+        document={document}
+      />
+
+      {props.doc.status === 'completed' &&
+      !props.doc.verified &&
+      !props.doc.waiting ? (
+        <DocumentProof
+          onClick={
+            props.user.credits <= 0
+              ? () => history.push('/buy')
+              : toggleVerifyPay
+          }
+        >
+          {props.doc.loading && !props.doc.error ? (
+            <LoadingContainer>
+              <BeatLoader color={'black'} sizeUnit={'px'} size={10} />
+            </LoadingContainer>
+          ) : props.doc.error ? (
+            'Error'
+          ) : (
+            'Click to Proof'
+          )}
+        </DocumentProof>
+      ) : (
+        <DocumentProof onClick={props.doc.verified ? toggleLinkModal : null}>
+          {props.doc.verified
+            ? 'See Proof'
+            : props.doc.waiting
+            ? 'Waiting...'
+            : 'Not signed'}
+        </DocumentProof>
+      )}
+      <ProofDocTextContainer>
+        <DocumentSubject onClick={getDocument}>
+          {cutDocSignTitle} <PreviewIcon className="fas fa-search" />
+        </DocumentSubject>
+        {/* Returns proofed timestamp if exists */}
+        {timestamp !== undefined ? (
+          <TimestampContainer>
+            Proofed
+            <Timestamp>
+              {moment
+                .utc(timestamp)
+                .local()
+                .format('D MMM YYYY hh:mma')}
+            </Timestamp>
+          </TimestampContainer>
+        ) : null}
+      </ProofDocTextContainer>
+    </DocumentContainer>
+  );
+};
 
 export default Document;
