@@ -1,21 +1,19 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import CTA from '../CTA/CTA.js';
 import LPcontent from '../LPcontent/LPcontent.js';
 import ReactGA from 'react-ga';
 
-class Home extends React.Component {
-  componentDidMount() {
+const Home = () => {
+  useEffect(() => {
     ReactGA.pageview('/');
-  }
+  }, []);
 
-  render() {
-    return (
-      <div className="App">
-        <CTA />
-        <LPcontent />
-      </div>
-    );
-  }
-}
+  return (
+    <div className="App">
+      <CTA />
+      <LPcontent />
+    </div>
+  );
+};
 
 export default Home;
