@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React, { useState } from 'react';
 import {
   Button,
   Modal,
@@ -15,19 +15,16 @@ const encode = data => {
     .join('&');
 };
 
-class PremiumModal extends Component {
-  state = {
-    name: '',
-    email: '',
-  };
+const PremiumModal = props => {
+  const [info, setInfo] = useState({ name: '', email: '' });
 
-  handleSubmit = e => {
+  const handleSubmit = e => {
     e.preventDefault();
 
     fetch('/', {
       method: 'POST',
       headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
-      body: encode({ 'form-name': 'contact', ...this.state }),
+      body: encode({ 'form-name': 'contact', ...info }),
     })
       .then(() => alert('Success. Thank you!'))
       .catch(error =>
@@ -36,69 +33,64 @@ class PremiumModal extends Component {
         )
       );
 
-    this.setState({
-      name: '',
-      email: '',
-    });
-
-    this.props.toggle();
+    setInfo({ name: '', email: '' });
+    props.toggle();
   };
 
-  handleChange = e => {
-    this.setState({ [e.target.name]: e.target.value });
+  const handleChange = e => {
+    setInfo({ ...info, [e.target.name]: e.target.value });
   };
 
-  render() {
-    const { name, email } = this.state;
-    const isEnabled =
-      name.length > 0 &&
-      email.length > 0 &&
-      email.includes('@') &&
-      email.includes('.');
+  const { name, email } = info;
 
-    return (
-      <Modal centered isOpen={this.props.isOpen} toggle={this.props.toggle}>
-        <ModalHeader toggle={this.props.toggle}>
-          Give us your email address, and we'll notify you when our premium
-          service is available.
-        </ModalHeader>
-        <ModalBody>
-          <Form>
-            <FormGroup>
-              <Input
-                type="text"
-                name="name"
-                id="name"
-                placeholder="Full Name"
-                value={name}
-                onChange={this.handleChange}
-              />
-            </FormGroup>
-            <FormGroup>
-              <Input
-                type="email"
-                name="email"
-                id="email"
-                placeholder="name@example.com"
-                value={email}
-                onChange={this.handleChange}
-              />
-            </FormGroup>
-          </Form>
-          <Button
-            disabled={!isEnabled}
-            size="lg"
-            block
-            color="info"
-            style={{ marginTop: '20px' }}
-            onClick={this.handleSubmit}
-          >
-            Sign Up
-          </Button>
-        </ModalBody>
-      </Modal>
-    );
-  }
-}
+  const isEnabled =
+    name.length > 0 &&
+    email.length > 0 &&
+    email.includes('@') &&
+    email.includes('.');
+
+  return (
+    <Modal centered isOpen={props.isOpen} toggle={props.toggle}>
+      <ModalHeader toggle={props.toggle}>
+        Give us your email address, and we'll notify you when our premium
+        service is available.
+      </ModalHeader>
+      <ModalBody>
+        <Form>
+          <FormGroup>
+            <Input
+              type="text"
+              name="name"
+              id="name"
+              placeholder="Full Name"
+              value={name}
+              onChange={handleChange}
+            />
+          </FormGroup>
+          <FormGroup>
+            <Input
+              type="email"
+              name="email"
+              id="email"
+              placeholder="name@example.com"
+              value={email}
+              onChange={handleChange}
+            />
+          </FormGroup>
+        </Form>
+        <Button
+          disabled={!isEnabled}
+          size="lg"
+          block
+          color="info"
+          style={{ marginTop: '20px' }}
+          onClick={handleSubmit}
+        >
+          Sign Up
+        </Button>
+      </ModalBody>
+    </Modal>
+  );
+};
 
 export default PremiumModal;
