@@ -1,12 +1,11 @@
-import React, { Fragment } from 'react';
-import { withRouter } from 'react-router';
+import React, { Fragment, useEffect } from 'react';
 import './App.css';
 import Home from './Home/Home.js';
 import axios from 'axios';
-import { connect } from 'react-redux';
+import { useSelector, useDispatch } from 'react-redux';
 import { getUserInfo } from '../actions/user';
 import { BeatLoader } from 'react-spinners';
-import { Switch, Route } from 'react-router-dom';
+import { Switch, Route, useLocation } from 'react-router-dom';
 import { HomeContainer } from './Home/HomeStyles.js';
 import styled from 'styled-components';
 import OurTeam from './OurTeam/OurTeam';
@@ -44,68 +43,85 @@ export const TwitterText = styled.p`
   font-size: 1rem;
 `;
 
-class App extends React.Component {
-  state = {
-    loading: true,
-  };
+function App() {
+  const user = useSelector(state => state.user.user);
+  const fetching = useSelector(state => state.user.retrieving);
+  const fetched = useSelector(state => state.user.retrieved);
+  const error = useSelector(state => state.user.error);
 
-  componentDidMount() {
-    this.props.getUserInfo();
-  }
+  const dispatch = useDispatch();
+  const { pathname } = useLocation();
 
-  render() {
+  useEffect(() => {
+    dispatch(getUserInfo());
+  }, []);
+
+  useEffect(() => {
     window.scrollTo(0, 0);
-    const user = this.props.user;
+  }, [pathname]);
 
-    if (this.props.fetching || (!this.props.fetched && !this.props.error)) {
-      return (
-        <LoadingContainer>
-          {/* Uncomment the TwitterAlert section if needed due to high traffic */}
-          {/* <TwitterAlert>
+  if (fetching || (!fetched && !error)) {
+    return (
+      <LoadingContainer>
+        {/* Uncomment the TwitterAlert section if needed due to high traffic */}
+        {/* <TwitterAlert>
             <TwitterText>
               Someone with <em>a lot</em> of Twitter followers tweeted about us,
               and we're experiencing abnormally high traffic. Thanks for your
               patience!
             </TwitterText>
           </TwitterAlert> */}
-          <BeatLoader color={'black'} loading={this.state.loading} />
-        </LoadingContainer>
-      );
-    }
-
-    return (
-      <Fragment>
-        <HomeContainer>
-          {user ? <DashboardNav /> : <TopNavBar />}
-          <Switch>
-            <Route exact path="/" component={user ? Documents : Home} />
-            <Route path="/team" component={OurTeam} />
-            <Route path="/privacy" component={Privacy} />
-            <Route path="/terms" component={Terms} />
-            {user && <Route path="/account" component={Billing} />}
-            {user && <Route path="/settings" component={Settings} />}
-            {user && <Route path="/buy" component={Buy} />}
-            <Route component={user ? Documents : Home} />
-          </Switch>
-        </HomeContainer>
-        <Footer />
-      </Fragment>
+        <BeatLoader color={'black'} />
+      </LoadingContainer>
     );
   }
+
+  return (
+    <Fragment>
+      <HomeContainer>
+        {user ? <DashboardNav /> : <TopNavBar />}
+
+        <Switch>
+          <Route exact path="/">
+            {user ? <Documents /> : <Home />}
+          </Route>
+
+          <Route path="/team">
+            <OurTeam />
+          </Route>
+
+          <Route path="/privacy">
+            <Privacy />
+          </Route>
+
+          <Route path="/terms">
+            <Terms />
+          </Route>
+
+          {user && (
+            <Route path="/account">
+              <Billing />
+            </Route>
+          )}
+
+          {user && (
+            <Route path="/settings">
+              <Settings />
+            </Route>
+          )}
+
+          {user && (
+            <Route path="/buy">
+              <Buy />
+            </Route>
+          )}
+
+          <Route>{user ? <Documents /> : <Home />}</Route>
+        </Switch>
+      </HomeContainer>
+      <Footer />
+    </Fragment>
+  );
 }
 
-const mapStateToProps = state => {
-  return {
-    user: state.user.user,
-    fetching: state.user.retrieving,
-    fetched: state.user.retrieved,
-    error: state.user.error,
-  };
-};
-
-export default withRouter(
-  connect(
-    mapStateToProps,
-    { getUserInfo }
-  )(App)
-);
+export default App;
